feat: 마이페이지 LNB 화면 구현
