+++ src/pages/home/index.tsx
@@ -2,36 +2,57 @@ import Banner from "./components/Banner/Banner";
 import { Wrapper } from "@/layouts/Layout";
 import styled from "styled-components";
 import LNB, { MenuItemsType } from "../../components/LNB/LNB";
-import Add from "@/assets/svg/home/Add";
 import PaginatedPromptSection from "./components/Prompt/PaginatedPromptSection";
 import Button from "@/components/common/Button/Button";
 import { useNavigate } from "react-router-dom";
 import Text from "@/components/common/Text/Text";
+import Icon from "@/components/common/Icon";
+import useToast from "@/hooks/useToast";
 
 export default function HomePage() {
     const navigate = useNavigate();
+    const showToast = useToast();
 
     const menuItems: MenuItemsType[] = [
         {
             key: "1",
             label: "텍스트 프롬프트",
             iconType: "TextBlock",
+            onClick: () => navigate("/"),
         },
         {
             key: "2",
             label: "이미지 프롬프트",
             iconType: "Image",
+            onClick: () =>
+                showToast(
+                    `이미지 프롬프트는 아직 준비 중인 기능이에요.`,
+                    "더 많은 프롬프트 탐색을 위해 빠르게 준비하고 있을게요!"
+                ),
+            disabled: true,
         },
         {
             key: "3",
             label: "동영상 프롬프트",
             iconType: "Video",
+            onClick: () =>
+                showToast(
+                    `동영상 프롬프트는 아직 준비 중인 기능이에요.`,
+                    "더 많은 프롬프트 탐색을 위해 빠르게 준비하고 있을게요!"
+                ),
+            disabled: true,
         },
         { type: "divider", key: "divider-1" },
         {
             key: "4",
             label: "저장한 프롬프트",
             iconType: "Bookmark",
+            onClick: () =>
+                showToast(
+                    `저장한 프롬프트는 아직 준비 중인 기능이에요.`,
+                    "더 많은 프롬프트 탐색을 위해 빠르게 준비하고 있을게요!"
+                ),
+            disabled: true,
         },
     ];
 
@@ -44,7 +65,7 @@ export default function HomePage() {
             onClick={handleClickNewButton}
             style={{ padding: "8px 12px", gap: 2 }}
         >
-            <Add />
+            <Icon name="Add" color="white" size={20} />
             <Text font="b2_16_semi" color="white">
                 프롬프트 등록
             </Text>
